+++ web/modules/custom/qs_activity/src/Form/EventEditForm.php
@@ -133,12 +133,22 @@ class EventEditForm extends EventEditFormBase {
 
     $form['body'] = [
       '#attributes' => ['required' => TRUE],
-      '#title' => $this->t('qs_activity.events.form.edit.body'),
-      '#placeholder' => $this->t('qs_activity.events.form.edit.body.placeholder'),
       '#type' => 'textarea',
       '#required' => FALSE,
       '#default_value' => $event->body->value,
     ];
+
+    $form['quill'] = [
+      '#markup' => '<div class="form-group">
+        <span class="quill-label">' . $this->t('qs_activity.events.form.edit.body') . '</span>
+        <div
+            id="editor-edit-event"
+            data-placeholder-translation="' . $this->t('qs_activity.events.form.edit.body.placeholder') . '"
+            class="quill-editor quill-editor-secondary form-textarea form-control">' . $event->body->value . '
+          </div>
+      </div>',
+    ];
+    $form['#attached']['library'][] = 'quartiers_solidaires/quill';
 
     $form['venue'] = [
       '#attributes' => [